requirements.txt
google-play-scraper
pandas
numpy
python-dateutil
transformers
torch
vaderSentiment
spacy
scikit-learn
psycopg2-binary
